project-root_1015/prediction-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pandas
numpy
joblib
tensorflow
python-dotenv
scikit-learn     
